Adds Primitive Assets
